# components/tenant-api/docs/swagger.yaml
basePath: /api/v1
definitions:
  config.EffectiveConfig:
    properties:
      defaults_chain:
        items:
          type: string
        type: array
      effective_config:
        additionalProperties: {}
        type: object
      merged_hash:
        type: string
      source_file:
        type: string
      source_hash:
        type: string
      tenant_id:
        type: string
      warnings:
        items:
          type: string
        type: array
    type: object
  config.ResolvedThreshold:
    properties:
      component:
        type: string
      customLabels:
        additionalProperties:
          type: string
        description: dimensional labels from {key="value"} syntax
        type: object
      metric:
        type: string
      regexLabels:
        additionalProperties:
          type: string
        description: regex labels from {key=~"pattern"} syntax
        type: object
      severity:
        type: string
      tenant:
        type: string
      value:
        format: float64
        type: number
    type: object
...
  internal_handler.PutTenantResponse:
    properties:
      message:
        type: string
      pr_number:
        type: integer
      pr_url:
        type: string
      status:
        type: string
      tenant_id:
        type: string
      warnings:
        items:
          type: string
        type: array
    type: object
  internal_handler.PutViewRequest:
    properties:
      description:
        maxLength: 1024
        type: string
      filters:
        additionalProperties:
          type: string
        type: object
      label:
        maxLength: 256
        minLength: 1
        type: string
    required:
    - filters
    - label
    type: object
  internal_handler.TenantDetail:
    properties:
      id:
        type: string
      raw_yaml:
        type: string
      resolved_thresholds:
        items:
          $ref: '#/definitions/config.ResolvedThreshold'
        type: array
      validation_warnings:
        items:
          type: string
        type: array
    type: object
...
  internal_handler.ViewResponse:
    properties:
      created_by:
        type: string
      description:
        type: string
      filters:
        additionalProperties:
          type: string
        type: object
      id:
        type: string
      label:
        type: string
    type: object
info:
  contact: {}
  description: |-
    Multi-tenant configuration management API for the Dynamic Alerting Platform.
    Provides CRUD over per-tenant alert thresholds, RBAC-scoped writes, and async batch operations.
    Auth is enforced by the oauth2-proxy sidecar and forwarded via X-Forwarded-Email / X-Forwarded-Groups.
  license:
    name: See repository LICENSE
  title: Tenant API
  version: v2.8.0
paths:
...
  /api/v1/tenants/{id}:
    get:
      description: Returns the raw YAML and resolved thresholds for a single tenant.
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.TenantDetail'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get tenant config
      tags:
      - tenants
    put:
      consumes:
      - application/yaml
      description: Validates, writes, and commits a tenant's YAML configuration.
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      - description: Tenant YAML content
        in: body
        name: body
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.PutTenantResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update tenant config
      tags:
      - tenants
...
  /api/v1/tenants/{id}/effective:
    get:
      description: |-
        Returns the tenant config after merging the _defaults.yaml
        chain from L0 (conf.d root) down to the tenant's own file.
        Includes two SHA-256 hashes (truncated to 16 hex chars):
        source_hash for raw file content and merged_hash for the
        canonical-JSON of the merged dict. Parity target:
        scripts/tools/dx/describe_tenant.py.
      parameters:
      - description: Tenant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/config.EffectiveConfig'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get tenant effective (merged) config
      tags:
      - tenants
...
  /api/v1/views/{id}:
    delete:
      parameters:
      - description: View ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a saved view
      tags:
      - views
    get:
      parameters:
      - description: View ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handler.ViewResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a single saved view
      tags:
      - views
    put:
      consumes:
      - application/json
      parameters:
      - description: View ID
        in: path
        name: id
        required: true
        type: string
      - description: View definition
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.PutViewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create or update a saved view
      tags:
      - views
swagger: "2.0"